# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/cameras/{cameraId}/zones": {
      "get": {
        "description": "Every rectangle configured on the camera, each with the alert level it raises. An empty list means the camera is either unmonitored or monitored full-frame.",
        "operationId": "ZonesController_findByCamera_v1",
        "parameters": [
          {
            "name": "cameraId",
            "required": true,
            "in": "path",
            "description": "Camera the zones belong to. Resolved inside the caller space only.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Zones of the camera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonitorZoneDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` — Caller has not completed their profile.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "FORBIDDEN"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` — No camera with that id in the caller space.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "NOT_FOUND"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the monitor zones of a camera",
        "tags": [
          "zones"
        ]
      },
      "post": {
        "description": "Admin only. Adds one percentage rectangle to a camera. Coordinates are percent of frame, not pixels, so the zone keeps its meaning if the recorder changes resolution. The rectangle must stay inside the frame and have a positive area; a rectangle that does not answers `INVALID_ZONE`.",
        "operationId": "ZonesController_create_v1",
        "parameters": [
          {
            "name": "cameraId",
            "required": true,
            "in": "path",
            "description": "Camera the zones belong to. Resolved inside the caller space only.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateZoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Zone created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorZoneDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body.<br>`INVALID_ZONE` — Rectangle is outside the frame or has no area.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "INVALID_ZONE"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` — Caller is not a space admin, or has an incomplete profile.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "FORBIDDEN"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` — No camera with that id in the caller space.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "NOT_FOUND"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a monitor zone",
        "tags": [
          "zones"
        ]
      }
    },
    "/api/v1/zones/{id}": {
      "get": {
        "description": "One rectangle with its alert level and the camera it belongs to.",
        "operationId": "ZonesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Zone id. Resolved inside the caller space only.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The zone.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorZoneDto"
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` — Caller has not completed their profile.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "FORBIDDEN"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` — No zone with that id in the caller space.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "NOT_FOUND"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read a monitor zone",
        "tags": [
          "zones"
        ]
      },
      "put": {
        "description": "Admin only. Partial update — the sent fields are merged onto the stored zone and the merged rectangle is what gets validated, so moving one edge cannot push the zone out of the frame.",
        "operationId": "ZonesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Zone id. Resolved inside the caller space only.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateZoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Zone after the merge.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorZoneDto"
                }
              }
            }
          },
          "400": {
            "description": "`VALIDATION_ERROR` — Malformed body.<br>`INVALID_ZONE` — The merged rectangle is outside the frame or has no area.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "VALIDATION_ERROR",
                            "INVALID_ZONE"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "`UNAUTHORIZED` — Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` — Caller is not a space admin, or has an incomplete profile.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "FORBIDDEN"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` — No zone with that id in the caller space.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "NOT_FOUND"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a monitor zone",
        "tags": [
          "zones"
        ]
      },
      "delete": {
        "description": "Admin only. Logical delete: the zone leaves every read and stops raising alerts, the alerts it already raised stay in the history.",
        "operationId": "ZonesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Zone id. Resolved inside the caller space only.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Zone removed. Empty body."
          },
          "401": {
            "description": "`UNAUTHORIZED` — Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "UNAUTHORIZED"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "`FORBIDDEN` — Caller is not a space admin, or has an incomplete profile.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "FORBIDDEN"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "`NOT_FOUND` — No zone with that id in the caller space.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiErrorDto"
                    },
                    {
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "NOT_FOUND"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a monitor zone",
        "tags": [
          "zones"
        ]
      }
    }
  },
  "info": {
    "title": "Tu Seguridad API",
    "description": "Space-scoped video surveillance API. Every route outside `auth` and `health` resolves resources inside the caller space, taken from the bearer token, so an id from another space answers 404 rather than a result. Writes that change what the system watches are admin-only. Failures answer `{ statusCode, code, message }` — branch on `code`, not on the message.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404,
            "description": "HTTP status, the same value as the response status line."
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "INVALID_ZONE",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "UPSTREAM_ERROR",
              "UPSTREAM_TIMEOUT",
              "INTERNAL_ERROR"
            ],
            "example": "NOT_FOUND",
            "description": "Stable machine-readable code. Branch on this, not on the message."
          },
          "message": {
            "type": "string",
            "example": "Camera 7f3a not found",
            "description": "Human-readable reason. Never carries credentials, tokens or snapshot URLs."
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
...
      "MonitorZoneDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "cameraId": {
            "type": "string"
          },
          "x": {
            "type": "number",
            "example": 12.5,
            "description": "Percent of frame width"
          },
          "y": {
            "type": "number",
            "example": 20,
            "description": "Percent of frame height"
          },
          "width": {
            "type": "number",
            "example": 30
          },
          "height": {
            "type": "number",
            "example": 40
          },
          "alertType": {
            "type": "string",
            "enum": [
              "intruder",
              "suspicious"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "cameraId",
          "x",
          "y",
          "width",
          "height",
          "alertType",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateZoneDto": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "example": 12.5
          },
          "y": {
            "type": "number",
            "example": 20
          },
          "width": {
            "type": "number",
            "example": 30
          },
          "height": {
            "type": "number",
            "example": 40
          },
          "alertType": {
            "type": "string",
            "enum": [
              "intruder",
              "suspicious"
            ],
            "description": "Alert level this zone raises"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height",
          "alertType"
        ]
      },
      "UpdateZoneDto": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "example": 12.5
          },
          "y": {
            "type": "number",
            "example": 20
          },
          "width": {
            "type": "number",
            "example": 30
          },
          "height": {
            "type": "number",
            "example": 40
          },
          "alertType": {
            "type": "string",
            "enum": [
              "intruder",
              "suspicious"
            ],
            "description": "Alert level this zone raises"
          }
        }
      }
    }
  }
}